First end-to-end working version
This commit is being typed with a NeXT keyboard!

It's working end-to-end... sort of. I'm currently only
using one of the keys in the USB HID descriptor, which
is probably limiting my typing speed fairly significantly.

I'd like to be able to increase that count. I think it might
require some global state, though...